- split between integer and floating-point data types styles - test improvement - read me update

+++ src/test/java/it/firegloves/mempoi/domain/MempoiColumnTest.java
@@ -61,7 +61,7 @@ public class MempoiColumnTest {
 
     @Test
     public void column_BIGINT() throws NoSuchMethodException {
-        this.assertMempoiColumn("column_BIGINT", Types.BIGINT, EExportDataType.DOUBLE);
+        this.assertMempoiColumn("column_BIGINT", Types.BIGINT, EExportDataType.INT);
     }
 
     @Test
